# public/doce/swagger.json
{
  "openapi": "3.0.0",

  "info": {
    "title": "MVC API",
    "version": "1.0.0",
    "description": "Simple PHP MVC REST API"
  },

  "servers": [
    {
      "url": "http://localhost/mvc/public/api"

    }
  ],

  "paths": {

    "/product": {

      "get": {
        "summary": "Get all products",

        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },

      "post": {

        "summary": "Create product",

        "requestBody": {

          "required": true,

          "content": {

            "application/json": {

              "schema": {

                "type": "object",

                "properties": {

                  "name": {
                    "type": "string"
                  },

                  "price": {
                    "type": "number"
                  },

                  "description": {
                    "type": "string"
                  },

                  "qty": {
                    "type": "integer"
                  }

                }

              }

            }

          }

        },

        "responses": {

          "201": {
            "description": "Product created"
          }

        }

      }

    },

    "/product/{id}": {

      "get": {

        "summary": "Get single product",

        "parameters": [

          {
            "name": "id",
            "in": "path",
            "required": true,

            "schema": {
              "type": "integer"
            }

          }

        ],

        "responses": {

          "200": {
            "description": "Success"
          },

          "404": {
            "description": "Product not found"
          }

        }

      },

      "put": {

        "summary": "Update product",

        "parameters": [

          {
            "name": "id",
            "in": "path",
            "required": true,

            "schema": {
              "type": "integer"
            }

          }

        ],

        "responses": {

          "200": {
            "description": "Updated"
          }

        }

      },

      "delete": {

        "summary": "Delete product",

        "parameters": [

          {
            "name": "id",
            "in": "path",
            "required": true,

            "schema": {
              "type": "integer"
            }

          }

        ],

        "responses": {

          "200": {
            "description": "Deleted"
          }

        }

      }

    },

    "/auth/register": {
      "post": {
        "summary": "Register new user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully"
          },
          "422": {
            "description": "Validation errors"
          }
        }
      }
    },

    "/auth/login": {
      "post": {
        "summary": "Login user",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "500": {
            "description": "Login failed"
          }
        }
      }
    },

    "/auth/logout": {
      "post": {
        "summary": "Logout user",
        "tags": ["Auth"],
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        }
      }
    },

    "/home": {
      "get": {
        "summary": "Home page (requires login)",
        "tags": ["Home"],
        "responses": {
          "200": {
            "description": "Welcome to home page"
          },
          "401": {
            "description": "Unauthorized - not logged in"
          }
        }
      }
    }

  }

}